requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.5.1
apache-airflow-providers-common-io==1.5.1
apache-airflow-providers-common-sql==1.24.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.1
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.1
apache-airflow-providers-sqlite==4.0.1
apispec==6.8.1
argcomplete==3.6.1
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
babel==2.17.0
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.9
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
dlib @ file:///C:/Users/USER/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
dm-tree==0.1.8
dnspython==2.7.0
docstring_parser==0.16
efficientnet-pytorch==0.7.1
email_validator==2.2.0
etils==1.11.0
exceptiongroup==1.2.2
executing==2.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.8
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.68.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
HyperPyYAML==1.2.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
immutabledict==4.2.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
keras==3.7.0
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
limits==4.4.1
linkify-it-py==2.0.3
llvmlite==0.44.0
lockfile==0.12.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.0.2
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prison==0.2.1
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.4
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
setproctitle==1.3.5
simple-parsing==0.1.6
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tokenizers==0.21.0
toml==0.10.2
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
wget==3.2
wirerope==1.0.0
wrapt==1.17.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0